backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.12
effdet==0.4.1
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langchain==0.0.235
langsmith==0.0.7
layoutparser==0.3.4
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib==3.7.2
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.0
Pillow==10.0.0
portalocker==2.7.0
protobuf==4.23.4
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.11
pypandoc==1.11
pyparsing==3.0.9
pypdfium2==4.18.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywin32==306
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scipy==1.11.1
six==1.16.0
SQLAlchemy==2.0.19
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.8.1
unstructured-inference==0.5.5
urllib3==2.0.3
Werkzeug==2.3.6
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2